+++ NEWS.md
@@ -21,6 +21,7 @@
 * `estimate_cdf()`: Added option `johnson_method` to specify the formula which is used for determining cumulative failure probabilities.
 
 ### Parametric Models
+* `rank_regression()`: Implementation of one- and two-parametric exponential distribution (`'exponential'` and `'exponential2'`). 
 * `ml_estimation()`: New arguments `start_dist_params` (optional vector with initial values of the parameters) and `control` (enables access to argument `control` in `optim()`).
 * `loglik_profiling()`: New argument `wts`. 
 * `loglik_profiling()` is now an S3 generic. `loglik_profiling()` becomes `loglik_profiling.default()`. Added `loglik_profiling.wt_reliability_data()`.